[#192] refactor: coupon 리페키징
