+++ src/main/java/com/sparta/popupstore/domain/coupon/dto/response/CouponCreateResponseDto.java
@@ -1,6 +1,6 @@
-package com.sparta.popupstore.domain.promotionevent.dto.response;
+package com.sparta.popupstore.domain.coupon.dto.response;
 
-import com.sparta.popupstore.domain.promotionevent.entity.Coupon;
+import com.sparta.popupstore.domain.coupon.entity.Coupon;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
